+++ FinalProjekt/Core/PayoutEngine.cs
@@ -4,12 +4,12 @@ namespace FinalProjekt.Core;
 
 public static class PayoutEngine
 {
-    public static int GetLogPayout(int bet, double baseMultiplier)
+    public static long GetLogPayout(long bet, double baseMultiplier)
     {
         double scaling = 1.2; 
         double logBonus = Math.Log10(Math.Max(1, bet)) * scaling;
         double finalMultiplier = baseMultiplier + logBonus;
         
-        return (int)Math.Round(bet * finalMultiplier);
+        return (long)Math.Round(bet * finalMultiplier);
     }
 }
